using MultiVision instead of VisionThread + GoogleVision
We are no longer using multithreading. Instead, we put all images into a single request, and then send this request to Google Vision API where, with 1 API call, it reads receipt images (receipts should be present in separate image files)
TODO: implement a feature that enables the program to read multiple receipts in a single image.
